Massive Overhaul of the SerialIO Behaviour
This commit includes the overhaul of the SerialIO (now called SerialIOController). The SerialIOController now functions as a abstract class which gets used by its counterparts, SerialWriter and SerialReader. In order to enable seperation of the serial writer and reader these classes serve to fulfill this purpose. Also several name changes/movements were made to the some of the headers.